requirements.txt
PyJWT
flask
python-dotenv
requests
responses
